+++ lib/src/encoder/encodable.dart
@@ -1,11 +1,7 @@
 library toml.src.encoder.encodable;
 
 /// Interface for an object which can be encoded as a TOML value or table.
-abstract class TomlEncodableValue {
-  /// Constant default constructor to allow subclasses with `const`
-  /// constructors.
-  const TomlEncodableValue();
-
+abstract interface class TomlEncodableValue {
   /// Converts this object to an object which can natively be represented as
   /// a TOML value or table.
   ///
@@ -20,16 +16,18 @@ abstract class TomlEncodableValue {
 /// that can be encoded as a TOML key, can always also be encoded as a TOML
 /// value since keys are strings and strings are TOML values.
 ///
-/// An object can have to different representations depending on whether it is
-/// used as a key or a value. By default the object is converted using
-/// [toTomlValue] in both cases. You have to `extend` or `mixin` this interface
-/// in order to use the default implementation of [toTomlKey].
-abstract class TomlEncodableKey extends TomlEncodableValue {
-  /// Constant default constructor to allow subclasses with `const`
-  /// constructors.
-  const TomlEncodableKey();
-
+/// An object can have two different representations depending on whether it is
+/// used as a key or a value. Use [TomlEncodableKeyMixin] if the key
+/// representation is the same as the value representation.
+abstract interface class TomlEncodableKey extends TomlEncodableValue {
   /// Like [toTomlValue] but is invoked when the object is used as a key
   /// of a `Map` instead of as a value.
+  dynamic toTomlKey();
+}
+
+/// Mixin for [TomlEncodableKey]s whose key representation is the same as
+/// their value representation.
+mixin TomlEncodableKeyMixin implements TomlEncodableKey {
+  @override
   dynamic toTomlKey() => toTomlValue();
 }